requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# HTTP Requests and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# Reddit API
praw==7.7.1

# Google APIs
google-generativeai==0.3.2
google-cloud-translate==3.15.3

# Alternative Translation (fallback)
googletrans==4.0.0rc1

# Task Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# JSON and Data Processing
jsonschema==4.20.0

# Logging
colorlog==6.8.0

# Railway deployment
gunicorn==21.2.0
